react router added, app broken apart for readibility and routing
